Create a Container
Now, to the container? First, the basics:

1.A service is just a fancy name a computer science major made up to describe a useful object. A logger is a useful object, so it's a service. A mailer object, a database connection object and an object that talks to your coffee maker's API: all useful objects, all services.

2.A container is an object, but it's really just an associative array that holds all your service objects. You ask it for a service by some nickname, and it gives you back that object. And it has some other super-powers that we'll see later.

Got it? Great, create a {dollar_sign}container variable and set it to a new ContainerBuilder object.

Hello Mr Container! Later, we'll see why Mr Container is called a builder.

Working with it is simple: use set to put a service into it, and get to fetch that back later. Call set and pass it the string logger. That's the key for the service - it's like a nickname, and we could use anything we want.

TIP The standard is to use lowercase characters, numbers, underscores and periods. Some other characters are illegal and while service ids are case insensitive, using lower-cased characters is faster. Want details? See github.com/knpuniversity/symfony-journey/issues/5.

Then pass the {dollar_sign}logger object. Now, pass {dollar_sign}container to runApp instead of the logger and update its argument. To fetch the logger from the container, I'll say {dollar_sign}container->get() then the key - logger.

The logger service goes into the container with set, and it comes back out with get. No magic.

Test it out:

php dino_container/roar.php
Yep, still roaring.