Oh, and now order doesn't matter. Down near the bottom, create a new Definition and pass it the StreamHandler class. We can remove the Logger and StreamHandler use statements too, because in a second, we won't be referencing these directly anymore. Just like before, call setArguments, pass it an array, and put the one constructor argument - the log file path - inside of it. Finish it off with {dollar_sign}container->setDefinition(), passing the service nickname as the first argument, and the Definition next.
And now that the logger.stream_handler service is being set with a Definition, we can remove the original code that set it directly.

So even though the logger service needs the stream handler service, we can describe them in any order. When we eventually ask for the logger, Symfony will go make the logger.stream_handler service first, then pass it to logger. That's why it's called a dependency injector container: it helps you manage dependencies.

But like before, our 'app code' has no idea any of this is happening. So when we hit the script again, there's another log message:

php dino_container/roar.php
tail dino_container/dino.log