This Definition object is massively important to Symfony's container, and in the framework, they're built behind-the-scenes all over the place. I'll show you how in a bit.
Beyond using the class name and passing constructor arguments, there are a bunch of other things that you can train a Definition object to do. For example, what if you wanted the container to instantiate a service, but then also call a method on it before passing the service back?

Suppose that I want to log a message as soon as the logger is created - even before it's returned from the container. To do that, use addMethodCall(). The Logger class has a debug method on it - pass that as the first argument. This is equivalent to calling {dollar_sign}logger->debug(). Then pass the array of arguments - we have just one, the message we want to log: 'The logger just got started'

With this, the container will create the logger, call the debug() method on it, and then pass it back. Test it!

php dino_container/roar.php
tail dino_container/dino.log
Brilliant! Now let's get harder.