There are actually two ways to add handlers to Logger: via the constructor, like we're doing now, OR by calling a pushHandler() method. Let's see if we can create a second handler and hook it up with a method call.
Start by creating the new Definition - it'll be a stream again, so re-use the StreamHandler class. This handler will dump the output to the console. To do that, call setArguments() like before, but this time, we'll pass a single argument: php://stdout. Whoops, and I'll fix my wrong variable name. Now put it into the container with setDefinition() - we'll call it logger.std_out_logger

ea, the file is getting kind long and ugly. That'll be fixed soon.

To register this handler with the logger service, we could just add it to the second constructor argument. But to prove we've got this mastered, let's use {dollar_sign}loggerDefinition->addMethodCall() and tell it to call a pushHandler method. This takes one argument: the actual handler object. To pass in the service we just setup, create a new Reference and give it the service name: logger.std_out_logger.

Open up the Logger class so we can see what's happening. Inside this class, there's a public pushHandler() method that has one argument.

Our new definition code says: call pushHandler on the Logger object and pass it the service whose id is logger.std_out_logger. With any luck, log messages will dump into our file and get printed out to the screen.

Let's do it!

php dino_container/roar.php
tail dino_container/dino.log
Hey, there's one message! But inside dino.log, we see both. So why did only one message get printed to the screen? Actually, it's just a matter of order: we're calling the debug() method before pushing the second handler.