Creating service definitions in PHP is just one way to do things: you can configure this same stuff purely in Yaml or XML. And since Symfony uses Yaml files, let's use them too.
Up top, create a {dollar_sign}loader object - set it to new YamlFileLoader() from the DependencyInjection component. This takes two arguments: the {dollar_sign}container and a new FileLocator. That's a really simple object that says: 'Yo, look for files inside of a config directory'. To make it read a Yaml file, call load() and point it at a new file called services.yml.

This code tells the container to go find service definitions in this file.

Now create a config/ directory and put that services.yml in there. Our goal is to move all of this Definition stuff into that Yaml file. We'll start with the logger. I'll comment out all of the  stuff, but keep it as reference