We know the second argument is an array of objects. So in Yaml, we need to surround this with square brackets to make that an array.
This time, no errors!

php dino_container/roar.php
tail dino_container/dino.log
The console message is gone, but the log file gets it.

The big point is that you can create Definition objects by hand, OR use a config file to do that for you. When services.yml is loaded, it is creating those same Definition objects.

And as you'll see in a bit, if you want to get really advanced, you'll want to understand both ways.