Adding a Parameter in PHP
Now let's move the other handler. But this one is a little trickier: its argument has a PHP expression - __DIR__. That's trouble.

But hey, ignore it for now! Copy the service name and put it into services.yml. The order of services does not matter. Pass it the class and give it a single argument. This will not work, but I'll copy the __DIR__.'/dino.log in as the argument:

That's the basic idea, but since that __DIR__ stuff is PHP code, this won't work. But the solution is really nice.

The container holds more than services. It also has a simple key-value configuration system called parameters. In PHP, to add a parameter, just say {dollar_sign}container->setParameter() and invent a name. How about root_dir? And we'll set its value to __DIR__.
That doesn't do anything, but now we can use that root_dir parameter anywhere else when we're building the container.