Caching the Container
Here's the question: can we take all of this metadata about the container and cache it somehow? Absolutely - and the way it caches is incredible.

After compiling, create a new variable called  and set it to a new PhpDumper object. Pass the {dollar_sign}container to the dumper.

This guy is an expert at taking that metadata and caching it to a file. To do that, use the good ol' fashioned file_put_contents - pass it some new file path - how about cached_container.php and for the contents, call {dollar_sign}dumper->dump().

Let's see what this does! Run the script again:

php dino_container/roar.php
Now the cached_container.php file pops into existence. And it's awesome.

The Cached Container

Oh, so many good things to see. First, notice that this dumps a PHP class that extends Container.

That's actually the same base class as the ContainerBuilder we've been working with, and it houses the all-important get() function that fetches out services. In other words, this ProjectServiceContainer looks and acts just like the {dollar_sign}container we're using now.

Next, this has our two parameter values sitting on top. And if you call getParameter() to fetch one, it just uses this array:
142 lines  dino_container/cached_container.php
... lines 1 - 16
class ProjectServiceContainer extends Container{
  private static {dollar_sign}parameters = array(
    'root_dir' => '/Users/weaverryan/Sites/knp/knpu-repos/symfony-journey-to-center/dino_container',
    'logger_startup_message' => 'Logger just got started!!! ',
  );
... lines 23 - 100
  public function getParameter({dollar_sign}name){
    {dollar_sign}name = strtolower({dollar_sign}name);
    if (!(isset(self::{dollar_sign}parameters[{dollar_sign}name]) || array_key_exists({dollar_sign}name, self::{dollar_sign}parameters))) {
      throw new InvalidArgumentException(sprintf('The parameter %s must be defined.', {dollar_sign}name));
    }
    return self::{dollar_sign}parameters[{dollar_sign}name];
  }
... lines 111 - 140
}
And now, the most important thing to notice: for each of our three services, there's a concrete method that's called when we ask for that service

142 lines  dino_container/cached_container.php
... lines 1 - 53
    /**
     * Gets the 'logger' service.
... lines 56 - 60
     */
  protected function getLoggerService(){
    {dollar_sign}this->services['logger'] = {dollar_sign}instance = new \Monolog\Logger('main', array(0 => {dollar_sign}this->get('logger.stream_handler')));
    {dollar_sign}instance->pushHandler(->get('logger.std_out_handler'));
    {dollar_sign}instance->debug('Logger just got started!!! ');
    return {dollar_sign}instance;
  }
... line 71
  /**
    * Gets the 'logger.std_out_handler' service.
... lines 74 - 78
   */
  protected function getLogger_StdOutHandlerService(){
    return {dollar_sign}this->services['logger.std_out_handler'] = new \Monolog\Handler\StreamHandler('php://stdout');
  }
... line 84
    /**
  / * Gets the 'logger.stream_handler' service.
... lines 87 - 91
   */
  protected function getLogger_StreamHandlerService(){
    return {dollar_sign}this->services['logger.stream_handler'] = new \Monolog\Handler\StreamHandler('/Users/weaverryan/Sites/knp/knpu-repos/symfony-journey-to-center/dino_container/dino.log');
  }
... lines 97 - 142
Seriously, if you look at the get() function in the parent class, you'll find that calling {dollar_sign}container->get('logger.std_out_logger') will ultimately execute this getLogger_StdOutLoggerService() method.

And these methods use the exact PHP code we would write to instantiate these objects directly. We pass the container Definition objects, and it dumps the raw PHP code that those represent.

This is even more incredible when you look at the getLoggerService() method:

 142 lines  dino_container/cached_container.php
... lines 1 - 61
  protected function getLoggerService(){
    {dollar_sign}this->services['logger'] = {dollar_sign}instance = new \Monolog\Logger('main', array(0 => {dollar_sign}this->get('logger.stream_handler')));
    {dollar_sign}instance->pushHandler({dollar_sign}this->get('logger.std_out_handler'));
    {dollar_sign}instance->debug('Logger just got started!!! ');
    return {dollar_sign}instance;
  }
... lines 71 - 142
Look closely: it creates the new Logger object, passes main and then passes an array, with a call to {dollar_sign}this->get('logger.stream_handler') to fetch that service from itself - the container. The second arguments key in the Yaml file causes this.

Next, it has our two method calls: pushHandler() with {dollar_sign}this->get('logger.std_out_logger') and then a call to debug(). Everything we put into those Definitions are dumped into a real PHP file that contains the raw code we would've written anyways.

So, if we use this container class directly, then fetching objects out of it could not be faster. Let's do it!