We rock at building containers. So now let's see how it's build inside of Symfony.
Setting up app_dev.php for Debugging

To figure things out, let's jump straight to the code, starting with the app_dev.php front controller. We're going to add some var_dump statements to core classes, and for that to actually work, we need to make a few changes here. First, instead of loading bootstrap.php.cache, require autoload.php. Second, make sure this {dollar_sign}kernel->loadClassCache() line is commented out.

A copy of some really core classes in Symfony are stored in the cache directory for a little performance boost. These two changes turn that off so that if we var_dump somewhere, it'll definitely work.