The Cached Container
Back in buildContainer(), let's var_dump() the  so far to see what we've got:

 811 lines  vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Kernel.php
... lines 1 - 628
  protected function buildContainer(){
... lines 631 - 640
    {dollar_sign}container = {dollar_sign}this->getContainerBuilder();
    var_dump({dollar_sign}container);die;
... lines 643 - 652
  }
... lines 654 - 811
Ok, refresh! Hmm, it didn't hit my code. Why? Well, the container might already be cached, so it's not going through the building process. To force a build, you can delete the cached container file. But before you do that, I'll look inside - it's located at app/cache/dev/appDevDebugProjectContainer.php:

 4243 lines  app/cache/dev/appDevDebugProjectContainer.php
... lines 1 - 16
class appDevDebugProjectContainer extends Container{
  private static {dollar_sign}parameters = array(
    'kernel.root_dir' => '/Users/weaverryan/Sites/knp/knpu-repos/symfony-journey-to-center/app',
    'kernel.environment' => 'dev',
... lines 22 - 640
  );
... lines 642 - 3820
  /**
  * Gets the 'user_agent_subscriber' service.
... lines 3823 - 3828
  protected function getUserAgentSubscriberService(){
    return {dollar_sign}this->services['user_agent_subscriber'] = new \AppBundle\EventListener\UserAgentSubscriber({dollar_sign}this->get('logger'));
  }
... lines 3833 - 4241
}
It's a lot bigger and has a different class name, but this is just like our cached container: it has all the parameters on top, then a bunch of methods to create the services. Now go delete that file and refresh.

rm app/cache/dev/appDevDebugProjectContainer.php
Great: now we see the dumped container. I want you to notice a few things. First, there are no service definitions at all. But we do have the 9 parameters. And that's it - the container is basically empty so far.