We can dump the parameters too - ->getParameterBag()->all():
 811 lines  vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Kernel.php
... lines 1 - 628
  protected function buildContainer(){
... lines 631 - 644
    if (null !== {dollar_sign}cont = {dollar_sign}this->registerContainerConfiguration({dollar_sign}this->getContainerLoader({dollar_sign}container))) {
      {dollar_sign}container->merge({dollar_sign}cont);
     }
     var_dump({dollar_sign}container->getParameterBag()->all());die;
... lines 649 - 652
  }
... lines 654 - 811
This dumps out the kernel parameters from earlier plus the stuff from parameters.yml.

So even though the container is still almost empty, we've nearly reached the end. This empty-ish container is returned to initializeContainer() where it's compiled and then dumped:

 810 lines  vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Kernel.php
... lines 1 - 551
  protected function initializeContainer(){
... lines 554 - 557
    {dollar_sign}container = {dollar_sign}this->buildContainer();
    {dollar_sign}container->compile();
    {dollar_sign}this->dumpContainer({dollar_sign}cache, {dollar_sign}container, {dollar_sign}class, {dollar_sign}this->getContainerBaseClass());
... lines 561 - 572
  }
... lines 574 - 810
Before compiling, we only have 1 service. But we know from running container:debug that there are a lot of services when things finish. The secret is in the compile() function, which does two special things: process dependency injection extensions and run compiler passes. Those are up next.