Registering of Extension Classes
The answer lives in the bundle classes. Open up AppBundle:

 10 lines  src/AppBundle/AppBundle.php
... lines 1 - 4
use Symfony\Component\HttpKernel\Bundle\Bundle;
class AppBundle extends Bundle{
}
This is empty, but it extends Symfony's base Bundle class. The key method is getContainerExtension():

212 lines  vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Bundle/Bundle.php
... lines 1 - 28
abstract class Bundle extends ContainerAware implements BundleInterface{
... lines 31 - 71
  public function getContainerExtension(){
    if (null === {dollar_sign}this->extension) {
      {dollar_sign}class = {dollar_sign}this->getContainerExtensionClass();
      if (class_exists({dollar_sign}class)) {
        {dollar_sign}extension = new {dollar_sign}class();
... lines 78 - 88
        {dollar_sign}this->extension = {dollar_sign}extension;
      } else {
        {dollar_sign}this->extension = false;
      }
    }
    if ({dollar_sign}this->extension) {
      return {dollar_sign}this->extension;
    }
  }
... lines 99 - 212
When Symfony boots, it calls this method on each bundle looking for something called an Extension. This calls getContainerExtensionClass() and checks to see if that class exists. Move down to that method:

 212 lines  vendor/symfony/symfony/src/Symfony/Component/HttpKernel/Bundle/Bundle.php
... lines 1 - 204
  protected function getContainerExtensionClass(){
    {dollar_sign}basename = preg_replace('/Bundle$/', '', {dollar_sign}this->getName());
    return {dollar_sign}this->getNamespace().'\DependencyInjection\'.{dollar_sign}basename.'Extension';
  }
... lines 211 - 212
Ah, and here's the magic. To find this 'extension' class, it looks for a DependencyInjection directory and a class with the same name as the bundle, except replacing Bundle with Extension. For example, for AppBundle, it's looking for a DependencyInjection\AppExtension class. We don't have that.

Open up the TwigBundle class and double-click the directory tree at the top to move PhpStorm here. TwigBundle does have a DependencyInjection directory and a TwigExtension inside:

 157 lines  vendor/symfony/symfony/src/Symfony/Bundle/TwigBundle/DependencyInjection/TwigExtension.php
... lines 1 - 25
class TwigExtension extends Extension{
... lines 28 - 33
  public function load(array {dollar_sign}configs, ContainerBuilder {dollar_sign}container){
... lines 36 - 130
  }
... lines 132 - 155
}
So because this is here, it's automatically registered with the container. We may not know what an extension does yet, but we know how it's all setup.