Extensions Configuration
The second job is to read our configuration array and use that information to mutate the service definitions. We'll see code that does this shortly.

Most extensions will have two lines near the top that call getConfiguration() and processConfiguration:

 158 lines  vendor/symfony/symfony/src/Symfony/Bundle/TwigBundle/DependencyInjection/TwigExtension.php
... lines 1 - 33
  public function load(array {dollar_sign}configs, ContainerBuilder {dollar_sign}container){
... lines 36 - 52
    {dollar_sign}configuration = {dollar_sign}this->getConfiguration({dollar_sign}configs, {dollar_sign}container);
    {dollar_sign}config = {dollar_sign}this->processConfiguration({dollar_sign}configuration, {dollar_sign}configs);
... lines 56 - 131
  }
... lines 133 - 158
Next to every extension class, you'll find a class called Configuration:

 201 lines  vendor/symfony/symfony/src/Symfony/Bundle/TwigBundle/DependencyInjection/Configuration.php
... lines 1 - 22
class Configuration implements ConfigurationInterface{
... lines 25 - 199
}
Watch out, a meteor! Oh, never mind, it's just the awesome fact that if I mess up some configuration - like globals as globalsss in Yaml, we'll get a really nice error. That doesn't happen by accident, that system evolved these Configuration classes to make that happen.

This is probably one of the more bizarre classes you'll see: it builds a tree of valid configuration that can be used under this key. It adds a globals section, which says that the children are an array. It even has some stuff to validate and normalize what we put here:

 201 lines  vendor/symfony/symfony/src/Symfony/Bundle/TwigBundle/DependencyInjection/Configuration.php
... lines 1 - 104
  private function addGlobalsSection(ArrayNodeDefinition {dollar_sign}rootNode){
    {dollar_sign}rootNode
      ->fixXmlConfig('global')
      ->children()
        ->arrayNode('globals')
          ->normalizeKeys(false)
          ->useAttributeAsKey('key')
          ->example(array('foo' => '@bar', 'pi' => 3.14))
          ->prototype('array')
... lines 115 - 124
            ->beforeNormalization()
              ->ifTrue(function ({dollar_sign}v) {
                if (is_array({dollar_sign}v)) {
                  {dollar_sign}keys = array_keys({dollar_sign}v);
                  sort({dollar_sign}keys);
                  return {dollar_sign}keys !== array('id', 'type') && {dollar_sign}keys !== array('value');
                }
                return true;
              })
            ->then(function ({dollar_sign}v) { return array('value' => {dollar_sign}v); })
          ->end()
... lines 138 - 147
        ->end()
      ->end()
    ->end()
   ;
 }
... lines 153 - 201
These Configuration classes are tough to write, but pretty easy to read. And if you can't get something to configure correctly, opening up the right Configuration class might give you a hint.

Back in TwigExtension, let's dump  after calling processConfiguration():

 158 lines  vendor/symfony/symfony/src/Symfony/Bundle/TwigBundle/DependencyInjection/TwigExtension.php
... lines 1 - 33
  public function load(array {dollar_sign}configs, ContainerBuilder {dollar_sign}container){
... lines 36 - 51
    {dollar_sign}configuration = {dollar_sign}this->getConfiguration({dollar_sign}configs, {dollar_sign}container);
    {dollar_sign}config = {dollar_sign}this->processConfiguration({dollar_sign}configuration, {dollar_sign}configs);
    var_dump({dollar_sign}config);die;
... lines 56 - 131
  }
... lines 133 - 158
This dumps out a nice, normalized and validated version of our config, including keys we didn't have, with their default values.