Extensions Mutate Definitions
So finally, how is the globals key used? Scroll down to around line 90:

 157 lines  vendor/symfony/symfony/src/Symfony/Bundle/TwigBundle/DependencyInjection/TwigExtension.php
... lines 1 - 33
  public function load(array {dollar_sign}configs, ContainerBuilder {dollar_sign}container){
... lines 36 - 86
      {dollar_sign}def = {dollar_sign}container->getDefinition('twig');
      foreach ({dollar_sign}config['globals'] as {dollar_sign}key => {dollar_sign}global) {
... lines 90 - 92
       {dollar_sign}def->addMethodCall('addGlobal', array({dollar_sign}key, {dollar_sign}global['value']));
... line 94
      }
    }
... lines 97 - 130
  }
... lines 132 - 157
For most people, this code will look weird. But not us! If there are globals, it gets the twig Definition back out of the ContainerBuilder. This definition was added when it loaded twig.xml, and now we're going to tweak it. Just focus on the second part of the if: it calls {dollar_sign}def->addMethodCall() and passes it addGlobal and two arguments: our key from the config, and the value - petrero in this case.

If you read the Twig documentation, it tells you that if you want to add a global variable, you can call addGlobal on the Twig_Environment object. And that's exactly what this does. This type of stuff is super typical for extensions.

If you refresh without any debug code, we'll get a working page again. Now open up the cached container - app/cache/dev/appDevDebugProjectContainer.php and find the method that creates the twig service - getTwigService(). Make sure you spell that correctly:

 4244 lines  app/cache/dev/appDevDebugProjectContainer.php
... lines 1 - 3703
  protected function getTwigService(){
    {dollar_sign}this->services['twig'] = {dollar_sign}instance = new \Twig_Environment(->get('twig.loader'), array('debug' => true, 'strict_variables' => true, 'exception_controller' => 'twig.controller.exception:showAction', 'form_themes' => array(0 => 'form_div_layout.html.twig'), 'autoescape_service' => NULL, 'autoescape_service_method' => NULL, 'cache' => '/Users/petrero/Sites/knp/knpu-repos/symfony-journey-to-center/app/cache/dev/twig', 'charset' => 'UTF-8', 'paths' => array()));
... lines 3707 - 3725
    {dollar_sign}instance->addGlobal('twitter_username', 'petrero');
    return {dollar_sign}instance;
  }
... lines 3730 - 4244
Near the bottom, we see it: {dollar_sign}instance->addGlobal('twitter_username', 'petrero'). We passed in simple configuration, TwigExtension used that to mutate the twig Definition, and ultimately the dumped container is updated.

That's the power of the dependency injection extensions, and if it makes even a bit of sense, you're awesome.

Our Configuration Wins

Oh, and one more cool note. If I added a twig service to config.yml, would it override the one from TwigBundle? Actually yes: even though the extensions are called after loading these files, any parameters or services we add here win.