Calling compile() executes a group of functions called compiler passes. In fact, there's one called MergeExtensionConfigurationPass, and it's responsible for the Extension system we just looked at:
 60 lines  vendor/symfony/symfony/src/Symfony/Component/DependencyInjection/Compiler/MergeExtensionConfigurationPass.php
... lines 1 - 21
class MergeExtensionConfigurationPass implements CompilerPassInterface{
... lines 24 - 26
  public function process(ContainerBuilder {dollar_sign}container){
... lines 29 - 38
    foreach ({dollar_sign}container->getExtensions() as {dollar_sign}name => {dollar_sign}extension) {
... lines 40 - 49
      {dollar_sign}extension->load({dollar_sign}config, {dollar_sign}tmpContainer);
... lines 51 - 52
    }
... lines 54 - 57
  }
}
It loops over the extension objects and calls load() on each one. This is where most of the services come from.

But there's a bunch of other compiler passes, and most do small things. They're usually registered inside your bundle class - FrameworkBundle is a great example:

 100 lines  vendor/symfony/symfony/src/Symfony/Bundle/FrameworkBundle/FrameworkBundle.php
... lines 1 - 45
class FrameworkBundle extends Bundle {
... lines 48 - 64
  public function build(ContainerBuilder {dollar_sign}container){
... lines 67 - 72
    {dollar_sign}container->addCompilerPass(new RoutingResolverPass());
    {dollar_sign}container->addCompilerPass(new ProfilerPass());
... lines 75 - 76
    {dollar_sign}container->addCompilerPass(new RegisterListenersPass(), PassConfig::TYPE_BEFORE_REMOVING);
    {dollar_sign}container->addCompilerPass(new TemplatingPass());
... lines 79 - 97
  }
}
The build() method of every bundle is called early, and is used almost entirely just to add compiler passes. So what's the point of compiler passes? Why not just do any container modifications in the extension class?

The special thing about a compiler pass is that when it's called, the entire container has been built. So it's perfect when you need to tweak the container, but only once all of the service definitions are loaded.