Compiler Pass and Tags
Let's see an example. In our services.yml, we have one service, and it's an event subscriber. To tell Symfony this is an event subscriber, we had to add the tag: kernel.event_subscriber:

 12 lines  app/config/services.yml
... lines 1 - 5
services:
  user_agent_subscriber:
    class: AppBundle\EventListener\UserAgentSubscriber
    arguments: ['@logger']
    tags:
      - { name: kernel.event_subscriber }
So how does that work?

It's a compiler pass! And you can see it registered in FrameworkBundle, it's RegisterListenerPass:

 107 lines  vendor/symfony/symfony/src/Symfony/Component/EventDispatcher/DependencyInjection/RegisterListenersPass.php
... lines 1 - 19
class RegisterListenersPass implements CompilerPassInterface{
... lines 22 - 43
  public function __construct({dollar_sign}dispatcherService = 'event_dispatcher', {dollar_sign}listenerTag = 'kernel.event_listener', {dollar_sign}subscriberTag = 'kernel.event_subscriber'){
    {dollar_sign}this->dispatcherService = {dollar_sign}dispatcherService;
    {dollar_sign}this->listenerTag = {dollar_sign}listenerTag;
    {dollar_sign}this->subscriberTag = {dollar_sign}subscriberTag;
  }
... lines 50 - 105
}
The subscriberTag property is: kernel.event_subscriber. Near the bottom, it calls ->findTaggedServiceIds() and passes it that:

 107 lines  vendor/symfony/symfony/src/Symfony/Component/EventDispatcher/DependencyInjection/RegisterListenersPass.php
... lines 1 - 87
  foreach ({dollar_sign}container->findTaggedServiceIds({dollar_sign}this->subscriberTag) as {dollar_sign}id => {dollar_sign}attributes) {
    {dollar_sign}def = {dollar_sign}container->getDefinition({dollar_sign}id);
... lines 90 - 94
    {dollar_sign}class = {dollar_sign}def->getClass();
... lines 96 - 102
    {dollar_sign}definition->addMethodCall('addSubscriberService', array({dollar_sign}id, {dollar_sign}class));
  }
... lines 105 - 107
It's saying: give me all services tagged with kernel.event_subscriber. The  variable at the bottom is the Definition object for the event_dispatcher. And we use it to add a method call for addSubscriberService and pass it the service id and the class.

Let's go see this in the cached container. Refresh to get it back, then search for user_agent_subscriber:

 4244 lines  app/cache/dev/appDevDebugProjectContainer.php
... lines 1 - 1126
  protected function getDebug_EventDispatcherService(){
    {dollar_sign}this->services['debug.event_dispatcher'] = {dollar_sign}instance = new \Symfony\Component\HttpKernel\Debug\TraceableEventDispatcher(new \Symfony\Component\EventDispatcher\ContainerAwareEventDispatcher({dollar_sign}this), {dollar_sign}this->get('debug.stopwatch'), {dollar_sign}this->get('monolog.logger.event', ContainerInterface::NULL_ON_INVALID_REFERENCE));
... lines 1130 - 1136
     {dollar_sign}instance->addSubscriberService('user_agent_subscriber', 'AppBundle\EventListener\UserAgentSubscriber');
... lines 1138 - 1160
     return {dollar_sign}instance;
  }
... lines 1163 - 4244
There it is! It's calling the addSubscriberService method and passing the service id and class.

This is one of the most common jobs for a compiler pass. For example, there's another tag called form.type and this FormPass looks for all services tagged with that and does some container tweaking.

And there's a bunch more: like the compiler pass that checks for circular references. If service A depends on service B, which depends on service C, which depends on service A, you'll get a really clear exception. Then there are other passes which make micro-optimizations to speed the container up even more.